# backend/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI 电台 DJ — 提交文本，异步生成电台音频",
        "title": "Hanhan Radio API",
        "contact": {
            "name": "Hanhan Radio Team"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/episodes": {
            "get": {
                "description": "按创建时间倒序返回最近 episode 列表（排除失败的），最多 50 条。列表中的 segments 不含 segue 全文。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "episodes"
                ],
                "summary": "获取 episode 列表",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_handler.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/backend_internal_handler.EpisodeList"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "提交用户输入文本，后台异步生成电台音频，立即返回 episode ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "episodes"
                ],
                "summary": "创建 episode",
                "parameters": [
                    {
                        "description": "用户输入",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_internal_handler.CreateEpisodeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_handler.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hanhan-radio_backend_internal_manager.Episode"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_handler.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/episodes/{id}": {
            "get": {
                "description": "根据 ID 查询 episode 状态和结果。前端轮询此接口直到 status 为 done 或 failed。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "episodes"
                ],
                "summary": "查询单个 episode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/backend_internal_handler.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/hanhan-radio_backend_internal_manager.Episode"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "episode 不存在",
                        "schema": {
                            "$ref": "#/definitions/backend_internal_handler.ErrorEnvelope"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "backend_internal_handler.CreateEpisodeRequest": {
            "type": "object",
            "properties": {
                "prompt": {
                    "type": "string",
                    "example": "想听点轻松的"
                }
            }
        },
        "backend_internal_handler.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "backend_internal_handler.EpisodeList": {
            "type": "object",
            "properties": {
                "episodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/hanhan-radio_backend_internal_manager.Episode"
                    }
                }
            }
        },
        "backend_internal_handler.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 40000
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "prompt is required"
                }
            }
        },
        "hanhan-radio_backend_internal_manager.Episode": {
            "type": "object",
            "properties": {
                "audio_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "prompt": {
                    "type": "string"
                },
                "segments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/hanhan-radio_backend_internal_manager.SongSegmentItem"
                    }
                },
                "status": {
                    "$ref": "#/definitions/hanhan-radio_backend_internal_manager.EpisodeStatus"
                }
            }
        },
        "hanhan-radio_backend_internal_manager.EpisodeStatus": {
            "type": "string",
            "enum": [
                "pending",
                "processing",
                "done",
                "failed"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusProcessing",
                "StatusDone",
                "StatusFailed"
            ]
        },
        "hanhan-radio_backend_internal_manager.SongSegmentItem": {
            "type": "object",
            "properties": {
                "artist": {
                    "type": "string"
                },
                "segue": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}